requirements.txt
dj-database-url==0.5.0
django-simpleui==2021.6.1